backend/pyproject.toml
[project]
name = "benchforge-backend"
version = "0.1.0"
description = "FastAPI backend for BenchForge."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16,<2.0",
  "cryptography>=45.0,<46.0",
  "fastapi>=0.116,<1.0",
  "greenlet>=3.1,<4.0",
  "httpx>=0.28,<1.0",
  "pydantic-settings>=2.10,<3.0",
  "reportlab>=4.4,<5.0",
  "sqlalchemy>=2.0,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "uvicorn[standard]>=0.35,<1.0",
]

[dependency-groups]
dev = [
  "mypy>=1.17,<2.0",
  "pytest>=8.4,<9.0",
  "pytest-asyncio>=1.1,<2.0",
  "ruff>=0.12,<1.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
plugins = []

[tool.alembic]
script_location = "alembic"
